Build 0.0.733 Experimental

+++ src/main/java/me/piitex/renjava/gui/overlay/Overlay.java
@@ -2,7 +2,7 @@ package me.piitex.renjava.gui.overlay;
 
 public interface Overlay {
 
-    int x();
+    double x();
 
-    int y();
+    double y();
 }
